+++ components/HomePage/LastReleases.tsx
@@ -1,9 +1,9 @@
 import MangaDex from "@consumet/extensions/dist/providers/manga/mangadex";
 
 const LastReleases = async () => {
-  const res = await fetch("https://kitsu.io/api/edge/manga/6");
-  const data = await res.json();
-  console.log(data);
+  // const res = await fetch("https://kitsu.io/api/edge/manga/6");
+  // const data = await res.json();
+  // console.log(data);
   // const mangaDex = new MangaDex();
 
   //   const res = await mangaDex.fetchMangaInfo("d8a959f7-648e-4c8d-8f23-f1f3f8e129f3")
